+++ ui-tests/tests/jupyterlab_export_svg_as_png_extension.spec.ts
@@ -1,4 +1,4 @@
-import { expect, test } from '@jupyterlab/galata';
+import { expect, test } from './galata';
 
 /**
  * Don't load JupyterLab webpage before running the tests.